==> 03.00-HP_modeling.qmd <==
```{css, echo=F}
p {
  text-align: justify
}
```

# Modelling of a heat pump and data reconciliation

```{r, echo=F, results='hide',label='activate venv'}
#path <- file.path('./venv/bin/python')
library(reticulate)
#use_python(path)
use_virtualenv('./venv', required=FALSE)
```


```{python, echo=F, results='hide', label='load data'}
import matplotlib
matplotlib.use('Agg')
import codes_01_energy_demand.NR_functions as fct1
import codes_03_HP_modeling.Regression_Mirco.carnot_regression as fct2
from codes_01_energy_demand.NR_functions import WeatherClustering
import numpy as np; from numpy import matlib
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats
from sklearn.cluster import KMeans, SpectralClustering, AgglomerativeClustering
import os
import plotly.express as px
import plotly.graph_objects as go
from IPython.display import display, HTML
import amplpy
from amplpy import AMPL, DataFrame
import sys
import re
import glob
```

## Introduction

The main heating utility at EPFL is an advanced two-stage heat pump. The importance of this equipment leads to an in detail investigation of its functioning. Hence, the goal of this section is to choose an appropriate working fluid for the heat pump based on environmental and economic characteristics. To do so, first a preliminary model in VALI, a process simulation software, is created. A degree of freedom (DOF) is conducted to find the necessary number of measurements. The following procedure is applied to perform the DOF analysis: 

- The state variables of the flows
- The balance equations, often being energy balance (EB), mass balance (MB), pressure drop (P) 
- The model/performances equations, describing how the unit act on an engineering perspective.

The state variables of the flows are characterized by a minimum number of state variables given by the Gibbs phase rule, i.e. for a stream of N compounds, N+2 variables are sufficient to fully characterize it, with at least one extensive variable. From that, all other properties can be derived, provided the right constitutive equations. At the end, the found data is going to be reconciled to ensure its validity. This allows to ultimately provide economic and thermodynamic performances of the system.

For this analysis, we'll consider the steady-state operation of the system, i.e. no accumulation.


```{r double-stage, out.width='80%', fig.align='center', fig.cap='Vali model of the double-stage heat pump',echo=F}
#| label: fig-double-stage
knitr::include_graphics('Figures/double_stage.png')
```

## System DOF analysis

A simple preliminary analysis allows to determine the overall minimum DOF of the system, giving a lower bound of measurements necessary to make it solvable. Representing the system as a flowsheet allows to retrieve the number of variables and equations. The variables consists of all the material flows (9 flows times 3 state variables), the electricity flows (2 flows) and the heat flows (4 flows). The equations are one mass balance and one energy balance per unit in the system (8 units times 2 flows) , the isentropic efficiencies for both compressors (2 parameters) as well as the isenthalpic expansions in the valves (2 parameters). Finally, having 9 flows, all of them having 1 phase, and 8 units yields:

$n_s = n_v - n_e =  (27+2+4) - (8+8+2+2) = 13$

And thus $n_s = 13$, meaning that we need **at least** 13 specifications to characterize the whole system. 

### VALI tags propositions
For this exercise, we're provided with 5 sets of possible tag arrangement. To goal is to determine which one of those is possible based on the DOF analysis. As we can see in (@tbl-DOFprop), all arrangement meet the minimum measurement requirements. However, we're not certain that the local DOF is respected. Summing $n_s$ for each unit processs and substracting the link between them should allow to determine to total number of DOF of the system taking local DOF into account. Since all tag arrangement provide a different number of set points, the solution should be given by the arrangement with matching number of DOF.

|       | Option 1 | Option 2 | Option 3 | Option 4 | Option 5 |
|-------|----------|----------|----------|----------|----------|
|DOF    | 14       | 18       | 19       | 16       | 13       |
: Measurements of the possible arrangement {#tbl-DOFprop}

## Unit models

For the sake of completeness, we present here the general unit models composing the system. 

::: {.panel-tabset}
## Compressor
Stating the unit model for a single compressor with one inlet stream and one outlet stream yields:

| Type  | Variables                                      | #        |
|-------|------------------------------------------------|----------|
| $n_x$ | $(n_{in}+n_{out})*(n_c+2)$                     | 6        |
| $n_p$ | $\Delta P$                                     | 2        |
|       | $\dot{W}_{in}$                                 |          |
| $n_t$ | $\eta_{is}$                                    | 1        |
| $n_v$ |                                                | 9        |

| Type  | Equations                                      | #        |
|-------|------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                 | 1        |
| (EB)  | $\dot{W}_{in}= \dot{m}_{in}c_p(T_{out}-T_{in})$| 1        |
| (P)   | $\Delta P=cst$                                 | 1        |
| (M)   | $\dot{W}_{in} =\eta_{is}\dot{W}_{in,is}$       | 1        |
| $n_e$ |                                                | 4        |

## Evaporator
For the evaporator, we proceed the same way. We model the heat exchange of the lake as a constant heat transfer, allowing to simplify the equation statement since only our main flow has to be considered. In the general case, the fluid enters the evaporator with a vapor quality $x_{in}<1$, but based on the assumption that the entering flow is purely gaseous allows to further simplify the modelisation. 

| Type  | Variables                                      | #        |
|-------|------------------------------------------------|----------|
| $n_x$ | $(n_{in}+n_{out})*(n_c+2)$                     | 6        |
| $n_p$ | $\Delta P$                                     | 1        |
|       | $\dot{Q}_{in}$                                 | 1        |
| $n_t$ | $\eta_{is}$                                    | 1        |
| $n_v$ |                                                | 9        |

| Type  | Equations                                        | #        |
|-------|--------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   | 1        |
| (EB)  | $\dot{Q}_{in}= \dot{m}_{in}c_p(T_{evap}-T_{in})$ | 1        |
| (P)   | $\Delta P=0$                                     | 1        |
| (M)   | $\dot{Q}_{in}= UA\cdot \text{LMTD}$              | 1        |
| $n_e$ |                                                  | 4        |

## Condensor
The VALI model suggest to model the condensor in two part: one taking care of the change of temperature (sensible), the second one modelling the phase change (latent). Using the same logic as the evaporator, only the main flow is considered and the other flows are modeled as heat transfer. The unit model equation will be the same for both condensor, and only the heat transfer model will vary. The base variables are the following:

| Type  | Variables                                      | #        |
|-------|------------------------------------------------|----------|
| $n_x$ | $(n_{in}+n_{out})*(n_c+2)$                     | 6        |
| $n_p$ | $\Delta P$                                     | 1        |
|       | $\dot{Q}$                                      | 1        |
| $n_t$ |                                                | 0        |
| $n_v$ |                                                | 8        |

::: {.panel-tabset}

### Sensible condensor
The model equation for the sensible condensor is to assume a constant pressure during the heat transfer:

| Type  | Equations                                        | #        |
|-------|--------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   | 1        |
| (EB)  | $\dot{Q}_{in}= \dot{m}_{in}c_p(T_{out}-T_{in})$  | 1        |
| (P)   | $\Delta P=0$                                     | 1        |
| (M)   | $\dot{Q}_{in}= UA\cdot \text{LMTD}$              | 1        |
| $n_e$ |                                                  | 4        |


### Latent condensor
The model equation for the latent condensor is to assume a constant temperature during the heat transfer:

| Type  | Equations                                        | #        |
|-------|--------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   | 1        |
| (EB)  | $\dot{Q}_{in}= \dot{m}_{in}\cdot L$              | 1        |
| (P)   | $\Delta P=0$                                     | 1        |
| (M)   | $\dot{Q}_{in}= UA\cdot \text{LMTD}$              | 1        |
| $n_e$ |                                                  | 4        |

And thus $n_s = n_x-n_e = 2$, meaning that we need 2 specifications to characterize the latent condensor.
:::

## Expansion valve

Assuming the expansion valves are not going through a phase change allows to ignore the vapor quality for both valves. The model implies an isenthalpic pressure drop.

| Type  | Variables                                      | #        |
|-------|------------------------------------------------|----------|
| $n_x$ | $(n_{in}+n_{out})*(n_c+2)$                     | 6        |
| $n_p$ | $\Delta P$                                     | 1        |
| $n_t$ | -                                              | 0        |
| $n_v$ |                                                | 7        |

| Type  | Equations                                        | #        |
|-------|--------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   | 1        |
| (EB)  | $0 = \dot{m}_{in}c_p(T_{out}-T_{in})$            | 1        |
| (P)   | $\Delta P = cst$                                 | 1        |
| (M)   | $h_{in}=h_{out}$                                 | 1        |
| $n_e$ |                                                  | 4        |


## Flash tank

The flash tank involves a little bit more of complexity due to the multiple flows mixing. 

| Type  | Variables                                      | #        |
|-------|------------------------------------------------|----------|
| $n_x$ | $(n_{in}+n_{out})*(n_c+2)$                     | 12       |
| $n_p$ | $\Delta P_1$                                   | 2        |
| $n_p$ | $\Delta P_2$                                   |          |
| $n_t$ | -                                              | 0        |
| $n_v$ |                                                | 14       |

| Type  | Equations                                        | #        |
|-------|--------------------------------------------------|----------|
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   | 2        |
| (MB)  | $\dot{m}_{in} = \dot{m}_{out}$                   |          |
| (EB)  | $0 = \dot{m}_{in}c_p(T_{out}-T_{in})$            | 1        |
| (P)   | $\Delta P_1 = 0$                                 | 2        |
| (P)   | $\Delta P_2 = 0$                                 |          |
| (M)   |                                                  |          |
| $n_e$ |                                                  | 5        |

:::

## VALI Tags selection
Now that all the types of unit models are characterized, it should be possible to determine the minimum number of measurements needed to make the system fully determined. While it should be possible to analytically compute the required number of DOF by summing the units models and substracting the links between them, it's also possible to discard a some options by looking at redundancies/overspecifications, which will prevent VALI from running:

- `Option 2`: All values are specified in the following equation $\Delta T_{min}=T_6-T_{out,bis}$, which makes the model overspecified.

- `Option 3`: All values are specified in the following equation $Q_{cond,load}=Q_{cond,sens}+Q_{cond,lat}$, which makes the model overspecified.

- `Option 5`: All values are specified in the following equation $COP_{real}=\dfrac{Q_{cond,load}}{Q_{cond,load}-Q_{evap}}$, which makes the model overspecified.

- `Option 1`: For this option, it is a bit more difficult. Let's observe only the first stage as a subsystem, as in @fig-first-stage. 

```{r first-stage, out.width='80%', fig.align='center', fig.cap='First stage of the heat pump',echo=F}
#| label: fig-first-stage
knitr::include_graphics('Figures/first_stage.png')
```

Let's first remove the flow $hp_9$, as it is the same as $hp_1$. Now for each flow, 4 variables describe it, linked by 2 equations. Considering no leakage, the mass flow is also defined. Lastly, for flow number 3 we declare its isentropic value:

| Flow  | Variables  | # Variables |Equations  | # Equations associated |
|-------|-----------|----------|--------|------|
| $hp_1$ | $T_1,p_1,h_1$    | 3     | $h_1=h(T_1,p_1)$ | 1 |
|        |      $s_1$       | 1     | $s_1=s(T_1,p_1)$ | 1 |
| $hp_2$ | $T_2,p_2,h_2$    | 3     | $h_2=h(T_2p_2)$  | 1 |
|        |      $s_2$       | 1     | $s_2=s(T_2,p_2)$ | 1 |
| $hp_3$ | $T_3,p_3,h_3$    | 3     | $h_3=h(T_3,p_3)$ | 1 |
|        |      $s_3$       | 1     | $s_3=s(T_3,p_3)$ | 1 |
|        |   $h_{3,s}$      | 1     | $h_{3,s}=h(p_3,s_2)$ | 1 |
| $hp_8$ | $T_8,p_8,h_8$    | 3     | $h_8=h(T_8,p_8)$ | 1 |
|        |      $s_8$       | 1     | $s_8=s(T_8,p_8)$ | 1 |
| $Mass$ | $\dot{m}$| 1 | - | 0 |

We can proceed similarly for each unit:

| Unit  | Variables  | # Variables |Equations | # Equations associated|
|-------|--------|-------|--------|------------|
| Valve | -    | 0    | $h_8=h_1$  | 1 |
| Evaporator | $\Delta p$    | 1  | $\Delta p=p_2-p_1$    | 1 |
|            | $Q_{evap}$    | 1  | $Q_{evap}=\dot{m}(h_2-h_1)$     | 1 |
| Compressor | $W_{comp,1}$    | 1  | $W_{comp,1}=\dot{m}(h_3-h_2)$   | 1 |
|  | $\pi$    | 1   | $\pi=\frac{p_3}{p_2}$   | 1 |
|  | $\eta$    | 1   | $\eta=\frac{h_{3,s}-h_2}{h_3-h_2}$   | 1 |
| Flash | $Q_f$    | 1   |  $Q_f=\dot{m}(h_3-h_8)$  | 1 |

and finally we can add 2 energy balances:

| System  | Variables  | # Variables | Equations | # Equations associated |
|-------|---------|----------|--------|--------|
| Global  | $Q_{load},W_{comp,2}$ | 2 | $Q_{load}=W_{comp,1}+W_{comp,2}+Q_{evap}$ | 1 |
| Local   |- | 0 | $Q_{f}=W_{comp,1} + Q_{evap}$ | 1 |

Let's summarize everything:

|  $~$ | # Variables |  # Equations associated | DOF |
|-------|---------|----------|--------|
| Total  | 26 | 18 |  8 |

Therefore, 8 values should be fixed. In `Option 1`, only 5 variables ($\eta,\Delta p, Q_{evap}, W_{comp,2}$ and $Q_{load}$) are specified. There could be a way to specify the pressure in the flash tank (specify $p_8$) or the heat transferred to the second stage (specify $Q_f$)  from the rest of the system, but even with that, it would only increase  the number of specifications up to 7, which is still one missing. Therefore, the system in `Option 1` is underspecified and cannot be solved.


Finally, that only leaves us with one viable options, which is `Option 4`. Using that set of inputs indeed let VALI run and reconcile data. Now that the model is running with our default fluid, we can inquire for other working fluids. 

### Model adjustment to a second working fluid 

Let's choose a new working fluid from @fig-working-fluid
```{r working_fluid, out.width='100%', fig.align='center', fig.cap='Selected refrigerants for the study', echo=F, fig.show='hold'}
#| label: fig-working-fluid
knitr::include_graphics('Figures/working_fluid.png')

```

The modern challenge of working fluid selection resides in balancing good thermal properties and environmental friendliness. As the latter is one of the critical parameters for all the project, let's ensure to use a fluid with a small global warming potential. The choice restricts to either Ammonia or Propylene when looking at the GWP. Even though propylene has a higher lifetime, it is considered as a safer fluid. In fact, ammonia is a very toxic chemical, being corrosive both to skin and eyes @Ammonia.  To be sure that propylene is a good choice, its thermal properties are studied in detail @Propene.
Starting with the freezing point, it is at 87.96K, which is low enough to ensure that the propylene is not going to be solid in the heat pump. Its critical point is at 46 bar, which is high enough to avoid it, leading to good fluid stability within the heat pump. When working with propylene, it is important to pressurize it, to enable liquefication. Typical pressure values of propylene are around 2-9 bar. Hence, there is an increased risk when working with this gas. As for the latent heat of vaporization, it has a similar value than propane, i.e. 438.96 kJ/kg. Lastly, due to its high lifetime, it should be avoided to release propylene into the environment and its flammability should be considered when working with it.


## Data reconciliation

Now that the double stage heat pump model is working for both fluids, let's get precise measurements. Experts provided us some measurements at different ambient temperatures, for both fluids and for low (50°C) and medium (65°C) temperature supply. Some measurements are redundant with the model. This is very good for us. Since the measurements may have errors due to sensors accuracy, bad calibration or random errors, we have the ability to correct them via data reconciliation.

Let $y_i$ (for $i=1,...,N_{mes}$) be a variable of the model which can be measured. Let $y_i^*$ be the actual measurement of $y_i$, measured with a sensor that has a standard deviation of $\sigma_i$. Let $x_j$ be a variable of the model that wasn't measured. Let's write the set of all $y_i$ and all $x_j$ as vectors: $X,Y$. Finally, let's write the model equations as constrains: $F(X,Y)=0$. Reconciliation aims getting a set of solutions which minimzes the total square error between model and measurements, weighted by each sensor accuracy. Therefore, the problem can be written:

$$\min_{X,Y} \sum_{i=1}^{N_{mes}}\left( \dfrac{y_i-y_i^*}{\sigma_i} \right)^2$$
$$\text{s.t. } F(X,Y)=0$$


Vali is able to perform data reconciliation when a measurement file is provided. Therefore, reconciled files are generated for both fluids and for both temperature supplies.
<!--
*Hint*:
Note that the measurement file indicates where the reconciled file should be saved. For instance, for the *R-290_LT_1.txt* file, the location is defined by the command `ARCHIVE -T Reconciled/reconc_R-290_LT_1.txt`. Make sure to have a folder named *Reconciled* in the same folder as the Vali model. -->

## Performance and cost evaluation
The purpose of this section is to evaluate the performance of the double stage heat pump and the cost of the equipments. The performance is evaluated for low temperature supply ($50°C$) and high temperature supply ($65°C$) for both fluids. The goal is to perform a regression on the second law efficiency to model its evolution with the external ambient temperature. In the last section,  data was measured on the heat pump for both fluids and both temperature levels at different external temperatures. Data reconciliation was performed with the model in order to minimize the measurements error. These results will now be used to evaluate the heat pump performance at different temperatures, and fit these measurements on a regression model. The following model will be used:
$$ C_{carnot}=a\cdot T_{ext}^2+ b\cdot T_{ext}+c$$


with $T_{ext}$ the external ambient temperature. Regression is performed using AMPL, which allows to simultaneously get the equipment price. First, let's write the optimization problem for regression:
\begin{array}{l}
\min{\sum_t (C_{factor,1}(t)-C_{factor,2}(t))^2} \\
\text{ s.t.}\\
\end{array}

| Constraint type                            | Equation                                                                                                                   |
|--------------------------------------------|----------------------------------------------------------------------------------------------------------------------------|
| Carnot efficiencies                          |                                                                                                                   |
| Real carnot efficiency                           | $C_{factor,1}=\dfrac{Q_{cond}(t)}{W_{hp}(t)}\dfrac{T_{ln,cond}(t)-T_{evap}(t)}{T_{ln,cond}(t)}~\forall t \text{ where } W_{hp}(t)>0$                                                                          |
| Regression model                           | $C_{factor,1}=a\cdot T_{ext}^2+ b\cdot T_{ext}+c$                                                                    |
| Log mean temperature of the condenser | $T_{ln,cond}(t)=\dfrac{T_{EPFL,in}(t)-T_{EPFL,out}(t)}{\ln{\left(\dfrac{T_{EPFL,in}(t)}{T_{EPFL,out}(t)}\right)}}$ |
| Work supplied by compressors | $W_{hp}(t)=W_{comp,1}(t)+W_{comp,2}(t)$ |

Where $Q_{cond}$ is the total heat load to supply, $W_{comp,1}$ is the energy supplied by the first compressor and $W_{comp,2}$ is the energy supplied by the second compressor of the two-stage heat pump. $T_{EPFL,in}$ and $T_{EPFL,out}$ are the heat exchanger temperatures at inlet and outlet respectively on the heat supply side. $T_{evap}$ is the temperature at the evaporator (where heat is taken from the lake). Lastly, the factors $a$, $b$, and $c$ are the variables to determine by thsi regression. The equipment cost can be retrieved using the following equations:
<details>
<summary>Click to display all the constraints</summary>
| Constraint type                            | Equation                                                                                                                   |
|--------------------------------------------|----------------------------------------------------------------------------------------------------------------------------|
| Costs |               |
| First compressor cost | $C_{comp,1}=\bigg(10^{k_1+k_2\log_{10}(W_{comp,2}^{max})+k_3\left(\log_{10}(W_{comp,1}^{max})\right)^2}\bigg) \dfrac{I}{I_{ref}}f_{BM}$|
| Second compressor cost | $C_{comp,1}=\bigg(10^{k_1+k_2\log_{10}(W_{comp,2}^{max})+k_3\left(\log_{10}(W_{comp,1}^{max})\right)^2}\bigg) \dfrac{I}{I_{ref}}f_{BM}$|
| Evaporator cost | $C_{evap}=\bigg(10^{k_{1,h}+k_{2,h}\log_{10}(A_{evap})+k_{3,h}\left(\log_{10}(A_{evap})\right)^2}\bigg) \dfrac{I}{I_{ref}}f_{BM,h}$|
| Condenser cost | $C_{cond}=\bigg(10^{k_{1,h}+k_{2,h}\log_{10}(A_{cond})+k_{3,h}\left(\log_{10}(A_{cond})\right)^2}\bigg) \dfrac{I}{I_{ref}}f_{BM,h}$|
| Areas |                               |
| Evaporator area | $A_{evap}=\dfrac{Q_{evap}^{max}}{U_{water,ref}*\Delta T_{ln,evap}}$ |
| Condenser area | $A_{cond}=\dfrac{Q_{cond}^{max}}{U_{air,ref}*\Delta T_{ln,cond}}$ |
| Log mean temperature differences (LMTD) |                               |
| Evaporator LMTD | $\dfrac{(T_{lake,in}-T_{evap}^{max})-(T_{evap}^{max}-T_{lake,out})}{\ln \left(\dfrac{T_{lake,in}-T_{evap}^{max}}{T_{evap}^{max}-T_{lake,out}} \right)}$ |
| Condenser LMTD | $\dfrac{(T_{cond}^{max}-T_{EPFL,in}^{max})-(T_{cond}^{max}-T_{EPFL,out}^{max})}{\ln \left(\dfrac{T_{cond}^{max}-T_{EPFL,in}^{max}}{T_{cond}^{max}-T_{EPFL,out}^{max}} \right)}$ |
| Maximum values |                  |
| Time index at maximum load | $t_{max}=\arg \max_{t}\big(Q_{cond}(t)\big)$
| Evaporator temperature at maximum load |  $T_{evap}^{max}=T_{evap}(t_{max})$        |
| Condenser temperature at maximum load |  $T_{evap}^{max}=T_{cond}(t_{max})$        |
| Inlet EPFL temperature at maximum load |  $T_{evap}^{max}=T_{EPFL,in}(t_{max})$        |
| Outlet EPFL temperature at maximum load |  $T_{evap}^{max}=T_{EPFL,out}(t_{max})$        |
| Maximum load |  $Q_{cond}^{max}=\max{\big(Q_cond(t)\big)}$        |
| Evaporator heat at maximum load |  $Q_{evap}^{max}=Q_{evap}(t_{max})$        |
| First heat pump power at maximum load |  $W_{comp,1}^{max}=W_{comp,1}(t_{max})$        |
| Second heat pump power at maximum load |  $W_{comp,2}^{max}=W_{comp,2}(t_{max})$        |
</details>

The results are shown in @fig-regression_results
```{python, echo=F, eval=T,label= get_regression,results='hide'}
results, observed =fct2.get_reg()
PATH = os.path.join(os.path.dirname(os.getcwd()),"report-group-3", "codes_03_HP_modeling","Regression_Mirco","regression_results.csv")


results.to_csv(PATH,float_format="%.4f")
```

```{r table-widget, echo=F, eval=T, fig.cap= 'Results of the regression'}
#| label: fig-regression_results
regression_results <- read.csv("codes_03_HP_modeling/Regression_Mirco/regression_results.csv")

library(DT)

datatable(regression_results, options = list(pageLength = 5, scrollX=TRUE)            
)
```

Note that cinv1 is in CHF/y and cinv2 is in CHF/kW/y. All areas are in m$^2$ and all costs are in CHF/yr We consider that for the heat pump at low temperature, it is able to provide heat at 55°C and that for medium temperature, it is able to provide heat at 70°C. The regression for each fluid is shown in @fig-plot_regression
```{python, echo=F, eval=T,label= plot_regression,fig.cap='Regression fitting on the second law efficiencies'}
#| label: fig-plot_regression
T_ext_span=np.linspace(0,15,100)
carnot_span=pd.DataFrame(index=T_ext_span)
colors=['blue','brown','green','red']

for j in results.index.values:
    carnot_span[j]=results.loc[j]['a']*np.power(T_ext_span,2)+results.loc[j]['b']*T_ext_span+results.loc[j]['c']

fig = px.line(carnot_span, x=carnot_span.index, y=carnot_span.columns, title='Carnot efficiency function of ambient temperature',labels={'index': 'Ambient temperature [°C]', 'value':'Carnot efficiency','variable': 'Fluids'},color_discrete_sequence=colors).add_scatter(x=[5],y=[0.5],mode='markers',marker=dict(color='black',opacity=0),name='Observed data').add_trace(go.Scatter(x=[5],y=[0.5],name='Regression',line=dict(color='black'),marker=dict(size=0,opacity=0)))



count=0
for j in results.index.values:
    fig.add_scatter(x=observed['T_ext_'+j],y=observed['Carnot_'+j],mode='markers',marker=dict(color=colors[count]),showlegend=False,)
    count+=1

```

The better performing heat pumps are those with propylene as a working fluid, but they are also the most expansive. The investment costs $cinv1$ and $cinv2$, as well as the coefficients $a$,$b$,$c$  will be used for the next section in all double stage heat pump models, for the assessment of the whole EPFL energy system. We will let the model decide which heat pump to use, dependind on the objective function.
